Enh(grouping): BREAKING change aggregation options and improve custom aggregation support
Add gridOption treeAggregationFns, to add custom aggregations with support for both the column menu and saveState.
Refactor native aggregations to follow the same api as custom aggregations.

Remove columnDef option treeAggregation and replace with separate options treeAggregationType and treeAggregationLabel

Remove "custom" aggregation type, as it is redundent with specifying a custom
aggregation function.

Allow custom aggregation finalizers to omit setting the rendered string. Render the value (and label if provided)
if a rendered string is not provided.

Fix groupingShowCounts option

Breaking changes are:
- options previously specified on the treeAggregation object are now  separate options (see docs)
- aggregation entity type property may be undefined in the case of a unnamed custom aggregation function
- aggregation entity previously set the initial value to 0, it now begins as undefined

Closes #3302