Add 2 options: In CellNav, add option to highlight the entire row and in Selection add option to show/hide highlight on selection
We have had to customize a lot in the grid to achieve the enhancements
requested. We use the grid in a way that makes the row the item which is
actionable, not an individual cell. It would help us a lot if the
following option were added:
1. In the cellNav feature, add an option that will highlight the entire
row rather than the just the cell on the row that has focus. This will
act as a highlight bar that will indicate the row that we will act upon.
2. In The selection feature, add an option that does not apply the row
highlight on selection. This will help us by not causing confusion when
used in conjunction with the highlight bar. The check box in the header
will signal the row is selected.
Both changes are included in this pull request.