feat(core): Accessibility and keyboard support to the filter controls.
The filter controls can now be accessed using the tab keys and the
remove filter button automatically moves the focus to the input when it
is disabled.
They also support OSX Voice over reading.
Additionally all of the google accessibility audit rules pass.
The focus input now provides a default aria label on the input that can
be overridden in the filter settings.