Makes all of the grid menus accessible.
When the grid opens the focus is automatically set to the first element
in the list. When the menu closes the focus is automatically reutnred to
the menu button that opened that list except for when the column is
hidden. When the column is hidden the focus is set to the master grid
menu.

Also adds the ability to have screen reader only buttons in the menus.
The example of this is the 'close' menu item that only appears when the
focus is over it.

Updates e2e tests to reflect menu item addition.

The menu now has an additional item that is visible to screen readers
only and only becomes visible when focused. This adds one item to all of
the tests that are counting the number of elements that are 'displayed'.