Added "columnsFilters" feature
Columns filters is a feature that enables the user to add a filter
button to the column header. This button opens a popup with a filter
form.
This extends the regular filtering by ui-grid by enabling an "or" logic
to filter by.

Fixed a bug with service.clear

The clear function would have thrown an error when clearing a non array
filter.term property.

Added "dynamic selectOptions"

Now also clears predefined filters

Fixed the dynamic select change

Added feature tutorial