feature(validation): add validation feature
This feature will make validation on fields possible.
It will not use angularjs basic validation because it
wouldn't work when editing is over.

Instead it adds custom properties to field considered
as invalid.

It provides a basic way to display error messages
using html title attribute and a cooler one that
depends on bootstrap to show a popover.

Tests and docs are provided.