Fix For Windows 10 Exporter
This is a fix for the Windows 10 users. Instead of using `.match` we
use `.search` which runs a boolean value. I noticed you guys were
trying to get the exact version of IE and that isn’t going to work for
Windows 10 users. I instead use the boolean value later to see if we
are even in IE. the `navigator.msSaveBlob` is a function in IE 10+ and
Edge so knowing that I can call that first and IE9 might possibly not
blow up. I haven’t gotten to officially test IE9 yet. If anything, this
PR is strictly a proof of concept. Feel free to message me with any
questions!