Remove code triggering scroll on each menu select
Each time toggling column visibility with the top right menu the menu is scrolled to the top. This gets impossibly annoying when having 20+ menu items. Once this code (that is mentioned as a "quick fix") is removed it stops scrolling the menu to the top when showing/hiding a column.