chore(build): change contents saved to package.json
The same contents saved to bower.json were being saved to package.json, generating an invalid package.json since the main attribute was an array instead of a string. Now this has been changed, and the package.json generated for distribution contains the same contents of the package.json file in the source code.

Fixes #4743