# This is a combination of 8 commits. # The first commit's message is: Added the clearFocusAfterEdit option to the gridOptions. This options is used in the afterCellEdit event in the cellNav directive.
If you have only one editOnFocus column and you click outside the grid while editing. The focus remains on the field and the editOnFocus will not work on a single click. This setting will fix that problem.

# This is the 2nd commit message:

Added the row index to the GridRow and updated the types for the documentation

# This is the 3rd commit message:

Added the CLEAR_FOCUS_EVENT and added the check if the cellNav is at the end of grid or the begin to clear the focus.

The CLEAR_FOCUS_EVENT can be called and really clearing the focus of the current focused column

# This is the 4th commit message:

When editOnFocus is enabled and you reach the end or beginning of the grid the focus will be removed.

After sorting the index is reset for the rows.

And removed unused code.

Fixes issue #4606

# This is the 5th commit message:

Revert the doc changes

# This is the 6th commit message:

Changed the unit tests because I changed the behaviour.

# This is the 7th commit message:

Skipped the testcases because it's not clear what they should do

# This is the 8th commit message:

Removed some code duplication of myself