Option to override cell navigation defaults
Option to override default cell navigation

keyDownOverrides will force the viewPortKeyDown event
to be triggered if the specified key event was pressed
down, even if it is a default key command for the grid.
The handling of the event should then be done in the
viewPortKeyDown handler since the grid is no longer
handling the event.