Apply column visibility properly
This fix removes an unneeded additional function when setting
column visibility. Since both functions were of the same
priority level, it was possible that the visibility that
was passed in would be overwritten with the default value of
true under certain circumstances. There was really no need
for two functions, as this default behavior can be accomplished
in one.