Fixing notify text if cellFilter takes string literal
Calling getCellDisplayValue to get the correct filtered value, instead of getIntersectionValueFiltered, which did not return the correct value if the cellFIlter contained a string literal as an argument.  This was the only place that function was being used, so it can be removed.  It was probably carried forward from the ng-grid source, which had the same issue.