fix(cellNav): notifyText incorrect if cellFilter had string literal parameter (#5404)
* Fixing notify text if cellFilter takes string literal

Calling getCellDisplayValue to get the correct filtered value, instead of getIntersectionValueFiltered, which did not return the correct value if the cellFIlter contained a string literal as an argument.  This was the only place that function was being used, so it can be removed.  It was probably carried forward from the ng-grid source, which had the same issue.

* Remove getIntersectionValueFiltered

This function was probably carried forward from ng-grid, and does not properly handle cellFilters with string literal arguments.  It was only being used in one place, and was easily replaced with getCellDisplayValue, which does return the proper value.