feat(filter): Add rawTerm option to columnDef filter options
Adds the rawTerm option to the columnDef filter options, when enabled
it causes the rowSearcher to skip trimming and escaping the search term,
providing the raw value to the specified filter condition. This is
intended for use with custom search condition functions where pre-escaping
all regular expression characters or trimming whitespace is not desired.

Relates to #2475