feat(emptyBaseLayer): made emptyBaseLayer module to create grid background
Made a new feature called 'emptyBaseLayer'. This can be added to the
ui-grid by using ui-grid-empty-base-layer. This creates a fake
background for the ui-grid of continues rows, the rows are not editable
or clickable, and just act as a background.
There also is a new tutorial page '218 Empty Base Layer'