test(pagination): add tests for custom pagination
Since custom pagination allows for arbitrary pages of rows, we pick some
feature of the existing data to paginate by.

I've chosen the approximate number of pencil strokes to write the
letters on column 2 (with a line through 'Z' so that it gets four
strokes, and we can have two pages with 4 rows).