fix(core): fix external template loading for grid-filter, grid-footer-cell and grid-header-cell
Fixes external template loading for `grid-filter`, `grid-footer-cell` and `grid-header-cell`.

This issue occurs when trying to use `filterHeaderTemplate`, `footerCellTemplate` or `headerCellTemplate` on a column definition with an external template (e.g. `myPersonalTemplate.html` as a value).
In fact, Angular UI grid was not waiting for the template to be loaded before executing `$elmt.append()` function and the template was not displayed (empty template).
This fix waits for the external template to be loaded first.