feat(sorting): Default sorting
Default sorting allows data to always have an implicit ordering while yielding
to any user-provided sorts. Default sorts effectively add additional columns to
the sort at the lowest priorities, they cannot be removed and automatically
reapply if the user clears their sort. Since they are intended to represent the
default state of the data the default sorts are not displayed on the column
headers and are effectively invisible to the user.