fix(6017): Fixes ui-grid-viewport scroll issue in chrome v56 (#6125)
Chrome v56 introduced a new feature called "scroll anchoring" which causes the scroll position to

update if element content above the current scroll position has changed. Since ui-grid-viewport updates the margin of the first line whenever a user is scrolling, scroll

anchoring will cause it to get into a scroll loop and scroll indefinitely (i.e. scroll to the end of

the grid). Credit to @Joel-Kornbluh