fix (expandable) expand all opens disableRowExpandable rows (#6442)
* fix (expandable) expand all opens disableRowExpandable rows

Resolved issue where clicking the "expand all" button opens sub grids that have disableRowExpandable set to true.
This leads to seeing ugly empty sub grid white space.  The fix simply includes a check for each row to see if expansion is disabled or not.

* fix (expandable) update to tests to handle expansion only in enabled rows

Tests expected all rows to be expanded when "toggle all" and "expand all" function are called.  This is not correct if a row's expansion has been turned off.