Update GridRenderContainer.js
The change to using integer column widths means that we sometimes have some grid width left over where column widths are rounded down for asterisk or percentage widths, see #6371. The logic that then re-allocates the remaining widths currently only does this over columns with asterisk widths so I've updated the logic to check if there are any asterisk columns and, if not, check for any percentage columns as a fall back. Then use one or other to allocate the remaining width over.