Fix bug of space to the right of the last column (#6513)
The change to using integer column widths means that we sometimes have some grid width left over where column widths are rounded down for asterisk or percentage widths, see #6371. The logic that then re-allocates the remaining widths currently only does this over columns with asterisk widths.

I've updated the logic to check if there are any asterisk columns and, if not, check for any percentage columns as a fall back. Only if there are no asterisk column widths, all columns are percentage widths and the percentages add up to 100%, then the percentage columns will be used to allocate remaining space, or trim if needed.