fix(menus): streamline menu positioning
Menu width calculation was removed in #6588. Menu width needs to be
calculated in order to properly position the menu (if there is not
enough room to the left of the column). repositionMenu is now always
called in the 'menu-shown' event listener to ensure the width can be
properly calculated. The style attribute is removed from $elm in the
'menu-hidden' event listener to prevent the menu from appearing to slide
from the left or right when opening another column menu while one is
already open.

Menu animation speed was changed in #6588 to have different add/remove
speeds. 0.04s has been chosen as a happy medium, and the $timeout
duration has been changed to reflect the new transition duration (to
reduce delay when hiding the menu).

`position: relative` has been removed from .ui-grid-header-cell-row to
ensure consistent calculation of offsetParent across browsers.

Fixes #5396, #5990, #6085.