feat(concat.js): Rename ui-grid.base.js to ui-grid.core.js.
After going through the ui-grid documentation, I realized that core makes more sense as the proper
name for the main UI-Grid file and decided to rename it as such.

BREAKING CHANGE: ui-grid.base.js has been rename to ui-grid.core.js. Also, ui-grid.core.js no longer
contains most of the language files available to UI-Grid. It only contains english as that is the
default language of the grid. See tutorials for examples.