Fix(cellnav.js), do not trigger edit on undefined event (#6691)
* fix(Grid.js): ScrollIfNecessary does not account for scrollWidth correctly

Since scrollIfNecessary is called multiple times when enableCellEditOnFocus is true we need to make sure the scrollbarWidth and footerHeight is accounted for to not cause a loop.

fixes #6653

* Add check for gridCol not null

Make sure gridCol is not null before checking for enableCellEditOnFocus

* fix(Grid.js) Vertical scroll calculates height wrong with enableHorizontalScrollbar: NEVER

Use scrollbarHeight instead of scrollbarWidth for vertical scroll calculations. scrollbarHeight has the value 0 when enableHorizontalScrollbar is set to NEVER.

Round calculated boundary values as computed style may have decimal number which will not match pixelsToSeeRow

* Fix(cellnav.js), do not trigger edit on undefined event

Do not trigger a cell edit when the event is undefined, if needed through API pass a null object instead.