fix(exporter): remove extra double quotes in excel export
Remove code that is adding a duplicate double quote for each double quote it sees in a string value when exporting from excel. Also change test file to reflect this change. The motivation is that it does not seem like normal behavior to add extra double quotes for excel exports and these extra double quotes become confusing for the end user.