fix: clearSelectedRows should respect enableSelection
The issue we had was that we have rows that are selected, but also have enableSelection: false. We expected select
all to behave similarly when selecting and unselecting all rows in that select all does not touch disabled rows. But
unselect all does touch disabled rows.

This fixes the above issue by using the same check that is used during select all behavior in clearSelectedRows.