Update selection.js
I think is not necessary know if the row is selectable or not, I'm having on issue related to it. Imagine you have a select dropdown with actions to perform to the rows selected, you select some rows but then you change the action to perform and this new action is not allowed to some of the rows previously selected, the function is not going to clear up those rows because they are selected and then they are not selectables.